Refactor: Consolidate duplicate make_test_participant into shared test_utils module
- Created tests/fpl_tests/test_utils.py with shared make_test_participant helper
- Removed duplicate implementations from test_chart_generator.py and test_statistics.py
- Added __init__.py files to tests/ and tests/fpl_tests/ for proper package structure
- All 71 tests pass successfully
