+++ tests/fpl_tests/__init__.py
@@ -1,0 +1,1 @@
+"""FPL tests package."""
