+++ tests/fpl_tests/test_chart_generator.py
@@ -1,26 +1,8 @@
 """Tests for chart generator functionality."""
 
 from fpl.chart_generator import generate_rank_progression_chart
-from fpl.participant import Participant
 
-
-def make_test_participant(
-    first_name: str,
-    history: list[dict],
-    entry_id: int = 1,
-    league_rank: int = 1
-) -> Participant:
-    """Helper to create test Participant objects."""
-    return Participant(
-        entry_id=entry_id,
-        team_name=f"{first_name}'s Team",
-        manager_name=f"{first_name} Doe",
-        total_score=0,
-        history=history,
-        last_event={},
-        league_rank=league_rank
-    )
-
+from .test_utils import make_test_participant
 
 
 def test_empty_chart_has_proper_axes():
